:hammer: Extract cross-platform logic into BaseSyncRouting for iOS/Android reuse
